requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.5.3
asgiref==3.11.0
astroid==3.3.11
attrs==25.3.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
certifi==2025.6.15
cffi==2.0.0
channels==4.2.0
channels-redis==4.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.4
daphne==4.1.2
dill==0.4.1
distro==1.9.0
Django==6.0.1
dnspython==2.7.0
exceptiongroup==1.3.0
filelock==3.18.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
isort==6.1.0
Jinja2==3.1.6
jiter==0.10.0
kombu==5.6.2
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.1.2
packaging==25.0
parsy==1.1.0
pathspec==0.12.1
platformdirs==4.9.2
playwright==1.57.0
presentations_module @ file:///Users/riskyworks/Documents/work/presentations/presentations-module
prompt_toolkit==3.0.52
protobuf==6.31.1
psycopg==3.2.10
psycopg-binary==3.2.10
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
pymongo==4.16.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.2
redis==4.6.0
regex==2024.11.6
requests==2.32.4
service-identity==24.2.0
setuptools==80.10.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
toml==0.10.2
tomlkit==0.14.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.5.0
zope.interface==8.2
